refactor(api): simplify error handling & testing
