+++ packages/api/package.json
@@ -12,7 +12,7 @@
     "start": "node-dev -r esm src/cli.js",
     "prepare": "babel --out-dir=dist --copy-files --source-maps=true src",
     "build": "npm run prepare",
-    "test": "jest --env node --no-cache src",
+    "test": "jest --env node --coverage --no-cache src",
     "postgres": "docker run --rm --name postgres -p 5432:5432 -v \"$(pwd)/data/postgres:/var/lib/postgresql/data\" postgres",
     "docker": "run-s docker:build docker:push",
     "docker:build": "docker build -t iameli/livepeer-api-prototype .",
@@ -23,6 +23,7 @@
     "any-base": "^1.1.0",
     "body-parser": "^1.18.3",
     "express": "^4.16.4",
+    "express-async-errors": "^3.1.1",
     "fs-extra": "^7.0.1",
     "level": "^5.0.1",
     "morgan": "^1.9.1",
@@ -38,7 +39,8 @@
     "esm": "^3.2.22",
     "jest": "^24.7.1",
     "node-dev": "^3.1.3",
-    "npm-run-all": "^4.1.5"
+    "npm-run-all": "^4.1.5",
+    "nyc": "^14.0.0"
   },
   "homepage": "https://livepeer.org",
   "repository": {
