+++ packages/api/src/store/errors.js
@@ -2,5 +2,6 @@ export class NotFoundError extends Error {
   constructor(message) {
     super(message)
     this.type = 'NotFoundError'
+    this.status = 404
   }
 }
